@@ -8,23 +8,22 @@ build-backend = "hatchling.build"
 name = "adjunct"
 dynamic = ["version"]
 description = "A collection of miscellaneous modules"
-license = { file = "LICENSE" }
-authors = [{ name = "Keith Gaughan", email = "keith@gaughan.ie" }]
-requires-python = ">=3.8"
 readme = "README.rst"
-keywords = []
+requires-python = ">=3.11"
+license = "MIT"
+license-files = ["LICENSE"]
+authors = [{ name = "Keith Gaughan", email = "keith@gaughan.ie" }]
 
 [project.urls]
 Documentation = "https://github.com/kgaughan/adjunct#readme"
 Issues = "https://github.com/kgaughan/adjunct/issues"
 Source = "https://github.com/kgaughan/adjunct"
 
-[tool.uv]
-dev-dependencies = ["mypy>=1.10.1", "pytest>=8.2.2", "pytest-cov>=2.12.1"]
-managed = true
+[dependency-groups]
+dev = ["mypy>=1.10.1", "pytest>=8.2.2", "pytest-cov>=2.12.1"]
 
 [tool.hatch.build.targets.sdist]
-include = ["/src/adjunct", "/requirements.lock"]
+include = ["src/adjunct", "uv.lock"]
 
 [tool.hatch.build.targets.wheel]
 packages = ["src/adjunct"]
@@ -39,34 +38,50 @@ version-file = "src/adjunct/_version.py"
 line-length = 120
 
 [tool.ruff]
-target-version = "py39"
+target-version = "py311"
 line-length = 120
-src = ["src", "test"]
+src = ["src", "tests"]
 
 [tool.ruff.lint]
 select = [
   "A",
+  # "ANN",
   "ARG",
   "B",
   "C",
   "DTZ",
   "E",
   "EM",
+  "EXE",
   "F",
   "FBT",
+  "FLY",
+  "FURB",
+  "G",
   "I",
   "ICN",
   "ISC002",
+  "LOG",
   "N",
+  "PERF",
+  "PIE",
   "PLC",
   "PLE",
   "PLR",
   "PLW",
+  # "PT",
+  "PYI",
   "Q",
+  "RET",
+  "RSE",
   "RUF",
   "S",
+  "SIM",
+  "SLOT",
   "T",
+  "TC",
   "TID",
+  "TRY",
   "UP",
   "W",
   "YTT",
@@ -82,6 +97,7 @@ ignore = [
   "UP035",   # Deprecated import
   "EM101",   # Exceptions must not use strings
   "EM102",   # Exception must not use f-string literal
+  "TRY003",  # Avoid specifying long messages outside the exception class
 ]
 
 [tool.ruff.lint.isort]
@@ -92,13 +108,15 @@ ban-relative-imports = "all"
 
 [tool.ruff.lint.per-file-ignores]
 # Tests can use magic values, assertions, and relative imports
-"tests/**/*" = ["PLR2004", "S101", "TID252", "E501"]
+"tests/**/*" = ["ANN", "PLR2004", "S101", "TID252", "E501"]
 
 [tool.pytest.ini_options]
 minversion = "6.0"
-addopts = "-ra -q --cov=adjunct --junitxml=tests/results.xml --cov-report html"
+addopts = "-ra -q --doctest-modules --cov=adjunct --junitxml=tests/results.xml --cov-report html"
 junit_logging = "out-err"
 junit_family = "xunit2"
+pythonpath = ["src"]
+testpaths = ["tests"]
 
 [tool.coverage.run]
 branch = true
